make the CloudWatch Logs group name configurable
Before, the group name was fixed, albeit with a configurable prefix.
This is a problem for existing systems that may want to adopt this
module, since they might already have flow logs going to a log group
with a different name. We want them to be able to use the same log
group.